% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po-object.R
\name{as_list_of_character}
\alias{as_list_of_character}
\title{Convert an object into a list of character vectors}
\usage{
as_list_of_character(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{An object.}

\item{.xname}{Not intended to be used directly.}
}
\value{
A list of character vectors.
}
\description{
Converts an object into a list of character vectors.  Not intended to be
called directly by the user, but needs to be exported to enable run-time
checking of \code{po} objects.
}

