% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_plural_forms.R
\name{get_n_plural_forms}
\alias{get_n_plural_forms}
\title{Get the number of plural forms from a PO file}
\usage{
get_n_plural_forms(x, default = 2L, ...)
}
\arguments{
\item{x}{A \code{po} object or its metadata element.}

\item{default}{An integer to return if the number of plural forms cannot be
determined.}

\item{...}{Arguments passed between methods.}
}
\value{
An integer of the number of plural forms for the language defined in
  the PO file.
}
\description{
Gets the number of plural forms specified in the"Plural-Forms" metadata
element of a PO file.
}
\note{
POT files are not language-specific and don't have a "Plural-Forms"
  metadata element. By convention, they are considered to have 2 plural
  forms, since that is how many pulral forms there are in English.
}

