% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_po_file}
\alias{read_po_file}
\alias{read_pot_file}
\title{Read PO and POT files}
\usage{
read_po_file(po_file)

read_pot_file(pot_file = NULL)
}
\arguments{
\item{po_file}{A string giving a path to a PO file.}

\item{pot_file}{A string giving a path to a POT file, or \code{NULL} to
autogenerate it.}
}
\value{
An object of class \code{po}, which is a list with the following
components:
\describe{
\item{type}{Either "r" or "c".  Guessed from the file name.}
\item{metadata}{A data frame of file metadata with columns "name" and
"value".}
\item{direct}{A data frame of messages with a direct translation, with
columns "msgid" and "msgstr".}
\item{countable}{A data frame of messages where the translation depends upon
a countable value (as created by \code{ngettext}), with columns "msgid",
"msgid_plural" and "msgstr".  The latter column contains a list of character
vectors.}
}
}
\description{
Reads .PO and .POT translation files.
}

