% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.po}
\alias{summary.po}
\title{Summarize a PO object}
\usage{
\method{summary}{po}(x, ...)
}
\arguments{
\item{x}{An object of class \code{po}.}

\item{...}{Arguments passed to \code{\link[base]{print.data.frame}}.}
}
\value{
The \code{x} argument in invisibly returned, but the function is
mostly invoked for the side-effect of printing the \code{x} argument.
}
\description{
\code{\link[base]{summary}} method for PO objects. \code{summary} shows less
information than \code{print}.
}
\examples{
pot_file <- system.file("extdata/R-summerof69.pot", package = "poio")
summary(pot <- read_po(pot_file))
}

