% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_po_file}
\alias{write_po_file}
\title{Write a PO file}
\usage{
write_po_file(po, po_file = NULL, ...)
}
\arguments{
\item{po}{An object of class \code{po}.}

\item{po_file}{A path to the po_file to be written, or NULL to automatically
generate the path.}

\item{...}{PAssed between methods. Not currently used.}
}
\value{
The funcion is mostly invoked for the side-effect of writing a PO
file.  The lines that are written to file are also invisibly returned.
}
\description{
Writes and object of class \code{po} to a .po file.
}

