% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_po}
\alias{read_po}
\alias{read_pot}
\title{Read PO and POT files}
\usage{
read_po(po_file)

read_pot(pot_file = NULL)
}
\arguments{
\item{po_file}{A string giving a path to a PO file.}

\item{pot_file}{A string giving a path to a POT file, or \code{NULL} to
autogenerate it.}
}
\value{
An object of class \code{po}, which is a list with the following
components:
\describe{
\item{source_type}{Either "r" or "c", depending upon whether the messages
originated from R-level code, or C-level code.  Determined from the file
name.}
\item{file_type}{Either "po" or "pot", depending upon whether the messages
originated from a PO (language-specific) or POT (master translation) file.
Determined from the file name.}
\item{metadata}{A data frame of file metadata with columns "name" and
"value".}
\item{direct}{A data frame of messages with a direct translation, with
columns "msgid" and "msgstr".}
\item{countable}{A data frame of messages where the translation depends upon
a countable value (as created by \code{ngettext}), with columns "msgid",
"msgid_plural" and "msgstr".  The latter column contains a list of character
vectors.}
}
}
\description{
Reads .PO and .POT translation files.
}
\examples{
# TODO
}
\seealso{
\code{\link[tools]{xgettext}}
}

